Move past purchase adapter into separate file
