+++ src/core/requests.ts
@@ -2,7 +2,7 @@ import { Request } from 'groupby-api';
 import { QueryTimeAutocompleteConfig, QueryTimeProductSearchConfig } from 'sayt';
 import Autocomplete from './adapters/autocomplete';
 import Configuration from './adapters/configuration';
-import RecommendationsAdapter from './adapters/recommendations';
+import PastPurchaseAdapter from './adapters/pastPurchases';
 import SearchAdapter, { MAX_RECORDS } from './adapters/search';
 import AppConfig from './configuration';
 import Selectors from './selectors';
@@ -33,7 +33,7 @@ namespace Requests {
       request.sort = <any>SearchAdapter.requestSort(sort);
     }
     if (Configuration.shouldAddPastPurchaseBias(config)) {
-      request.biasing = RecommendationsAdapter.pastPurchaseBiasing(state);
+      request.biasing = PastPurchaseAdapter.pastPurchaseBiasing(state);
     }
 
     return <Request>Requests.chain(config.search.defaults, request, config.search.overrides);
